refactor(core)!: move media capability predicates to core layer
Move the media capability type guards and their sentinel constants from
the DOM layer (dom/media) into the runtime-agnostic core/media layer.
Replace the DOM-lib types in isQuerySelectorAllCapable with a
framework-agnostic NodeListLike look-alike so it compiles under the core
project's WebWorker lib.

BREAKING CHANGE: the media capability guards (isMediaPauseCapable,
isMediaSeekCapable, etc.) and isQuerySelectorAllCapable now export from
@videojs/core instead of @videojs/core/dom.
