+++ packages/core/src/core/index.ts
@@ -1,3 +1,4 @@
+export * from './media/predicate';
 export * from './media/state';
 export * from './media/types';
 export * from './ui/airplay-button/airplay-button-core';
