+++ packages/core/src/dom/index.ts
@@ -8,7 +8,6 @@ export * from './hotkey/aria';
 export * from './hotkey/coordinator';
 export * from './hotkey/hotkey';
 export * from './hotkey/hotkey-events';
-export * from './media/predicate';
 export * from './media/types';
 export * from './store/features';
 export * from './store/selectors';
