+++ packages/core/src/dom/store/features/buffer.ts
@@ -1,8 +1,7 @@
 import { listen, serializeTimeRanges } from '@videojs/utils/dom';
-
+import { isMediaBufferCapable } from '../../../core/media/predicate';
 import type { MediaBufferState } from '../../../core/media/state';
 import { definePlayerFeature } from '../../feature';
-import { isMediaBufferCapable } from '../../media/predicate';
 
 export const bufferFeature = definePlayerFeature({
   name: 'buffer',
