+++ packages/core/src/dom/store/features/controls.ts
@@ -1,10 +1,9 @@
 import { listen } from '@videojs/utils/dom';
 import { isNull } from '@videojs/utils/predicate';
-
+import { isMediaPauseCapable, isMediaRemotePlaybackCapable } from '../../../core/media/predicate';
 import type { MediaControlsState } from '../../../core/media/state';
 import { definePlayerFeature } from '../../feature';
 import { findGestureCoordinator } from '../../gesture/coordinator';
-import { isMediaPauseCapable, isMediaRemotePlaybackCapable } from '../../media/predicate';
 import { isRemotePlaybackConnected, isRemotePlaybackConnecting } from '../../presentation/remote-playback';
 
 const IDLE_DELAY = 2000;
