+++ packages/core/src/dom/store/features/quality.ts
@@ -1,9 +1,8 @@
 import { listen } from '@videojs/utils/dom';
-
+import { isMediaVideoDimensionsCapable, isMediaVideoRenditionCapable } from '../../../core/media/predicate';
 import type { MediaQualityState, MediaVideoRendition } from '../../../core/media/state';
 import type { VideoRenditionLike, VideoRenditionListLike } from '../../../core/media/types';
 import { definePlayerFeature } from '../../feature';
-import { isMediaVideoDimensionsCapable, isMediaVideoRenditionCapable } from '../../media/predicate';
 
 const QUALITY_AUTO_VALUE = 'auto';
