+++ packages/core/src/dom/store/features/time.ts
@@ -1,8 +1,13 @@
 import { listen, onEvent } from '@videojs/utils/dom';
 import { noop } from '@videojs/utils/function';
+import {
+  hasMetadata,
+  isMediaBufferCapable,
+  isMediaSeekCapable,
+  isMediaSourceCapable,
+} from '../../../core/media/predicate';
 import type { MediaTimeState } from '../../../core/media/state';
 import { definePlayerFeature } from '../../feature';
-import { hasMetadata, isMediaBufferCapable, isMediaSeekCapable, isMediaSourceCapable } from '../../media/predicate';
 import { signalKeys } from '../signal-keys';
 
 export const timeFeature = definePlayerFeature({
